Fix theater level bacgrounds

+++ src/background.zig
@@ -5,7 +5,7 @@ const shared = @import("shared.zig");
 const camera = @import("camera.zig");
 
 pub const SerializableParallaxEntity = struct {
-    parallaxDistance: i32,
+    parallaxDistance: f32,
     fog: f32,
     imgPath: []const u8,
     scale: vec.Vec2,
@@ -13,7 +13,7 @@ pub const SerializableParallaxEntity = struct {
 };
 
 pub const ParallaxEntity = struct {
-    parallaxDistance: i32,
+    parallaxDistance: f32,
     fog: f32,
     scale: vec.Vec2,
     pos: vec.IVec2,
@@ -40,7 +40,7 @@ pub fn draw() !void {
     }
 }
 
-pub fn create(s: sprite.Sprite, pos: vec.IVec2, parallaxDistance: i32, scale: vec.Vec2, fog: f32) !void {
+pub fn create(s: sprite.Sprite, pos: vec.IVec2, parallaxDistance: f32, scale: vec.Vec2, fog: f32) !void {
     const parallaxEntity = ParallaxEntity{
         .sprite = s,
         .parallaxDistance = parallaxDistance,
